Alterado ids de Integer para Long e adicionado o Swagger

+++ backend-java/src/main/java/br/com/nemeia/brigia/service/MedicalPlanService.java
@@ -22,7 +22,7 @@ public class MedicalPlanService {
         return mapper.toResponseList(repository.findAll());
     }
 
-    public MedicalPlanResponse getById(Integer id) {
+    public MedicalPlanResponse getById(Long id) {
         return repository.findById(id)
                 .map(mapper::toResponse)
                 .orElseThrow(() -> new MedicalPlanNotFoundException("Plano não encontrado com ID: " + id));
